🔥 불필요한 코드 삭제
* issue: [NAILCASE-248](https://nailcase.atlassian.net/browse/NAILCASE-248)
sub-issues:
- [NAILCASE-348](https://nailcase.atlassian.net/browse/NAILCASE-348)
