+++ src/component/custom/customer/shop/05/index.tsx
@@ -1,15 +1,19 @@
 "use client"
 
-import { useRef } from "react"
+import { useRouter } from "next/navigation"
+import type { Dispatch, SetStateAction } from "react"
+import { useRef, useState } from "react"
 
+import { NTButton } from "@/component/common/atom/nt-button"
 import CustomerNaviBar from "@/component/custom/customer/shop/03"
 import ShopInfoCardList from "@/component/custom/customer/shop/04"
 import PostCardList from "@/component/custom/customer/shop/05/01"
 import ShopDesignList from "@/component/custom/customer/shop/05/02"
 import ShopNewsList from "@/component/custom/customer/shop/05/03"
 import ShopReviewList from "@/component/custom/customer/shop/05/04"
+import { getNowStamp } from "@/util/common"
 
-export default function CustomerShopContent() {
+export default function CustomerShopContent({ shopId }: { shopId: number }) {
 	const shopInfoRef = useRef<HTMLDivElement>(null)
 	const designRef = useRef<HTMLDivElement>(null)
 	const newsRef = useRef<HTMLDivElement>(null)
@@ -36,8 +40,12 @@ export default function CustomerShopContent() {
 					if (tool === "리뷰") return handleScroll(reviewRef)
 				}}
 			/>
-			<div ref={shopInfoRef} className="flex w-full flex-col gap-[20px] pt-16">
+			<div
+				ref={shopInfoRef}
+				className="flex w-full flex-col gap-[20px] pt-[32px]"
+			>
 				<ShopInfoCardList />
+				<ReservationBox shopId={shopId} />
 				<p className="text-Title02">네일샵 공지</p>
 				<PostCardList />
 			</div>
@@ -51,5 +59,32 @@ export default function CustomerShopContent() {
 				<ShopReviewList />
 			</div>
 		</div>
+	)
+}
+
+function ReservationBox({ shopId }: { shopId: number }) {
+	const [clickedStamp, setClickedStamp] = useState(getNowStamp())
+
+	return <ReservationButton {...{ clickedStamp, setClickedStamp, shopId }} />
+}
+type ReservationScheduleSubComponentPT = {
+	clickedStamp: number
+	setClickedStamp: Dispatch<SetStateAction<number>>
+	shopId: number
+}
+
+function ReservationButton({ shopId }: ReservationScheduleSubComponentPT) {
+	const router = useRouter()
+	return (
+		<NTButton
+			variant="tertiary"
+			flexible="full"
+			size="small"
+			onClick={() => {
+				router.push(`/shop/${shopId}/reservation`)
+			}}
+		>
+			예약하기
+		</NTButton>
 	)
 }
